requirements.txt
numpy >= 1.26.0
hydra-core >= 1.3
praw >= 7.7.0
rich >= 14.0.0
rally @ git+https://github.com/anton-pershin/rally.git
